fixes neural network

+++ common/common.h
@@ -34,7 +34,7 @@ typedef enum LOG_LEVEL {
 void list_init(list_t **lst, size_t count, void **data, free_fp ffp);
 size_t list_len(list_t *lst);
 void *list_get(list_t *lst, long index);
-void list_set(list_t *lst, size_t index, void *value);
+void list_set(list_t *lst, long index, void *value);
 void *list_set_get_existing(list_t *lst, size_t index, void *value);
 void list_swap(list_t *lst, size_t i, size_t j);
 void list_shuffle(list_t *lst);
